simplification fichiers -> utilisation items
