+++ apps/nuxt/src/composables/useBaserowPriorities.ts
@@ -1,4 +1,4 @@
-import type { BaserowProject, ProjectRow } from '~/types/baserow'
+import type { BaserowProject, ProjectRow, HistoryEntry } from '~/types/baserow'
 
 const getTheme = (project: BaserowProject): string => {
   const theme = project['Thématique principale']
@@ -9,13 +9,14 @@ const getTheme = (project: BaserowProject): string => {
 
 export const useBaserowPriorities = (selectedNaf: Ref<string>) => {
   const projects = ref<BaserowProject[]>([])
+  // Snapshot des valeurs avant modification — sert à filtrer les projets changés et à renseigner l'historique
+  const originalProjects = ref<Map<number, { Prio: number | string; 'Prios spécifiques': string }>>(new Map())
+  const history = ref<HistoryEntry[]>([])
   const loading = ref(false)
   const saving = ref(false)
   const successMessage = ref('')
   const errorMessage = ref('')
-  const modifiedIds = ref(new Set<number>())
-
-  const hasChanges = computed(() => modifiedIds.value.size > 0)
+  const hasChanges = ref(false)
 
   const getPriority = (project: BaserowProject): number => {
     const naf = selectedNaf.value
@@ -37,14 +38,27 @@ export const useBaserowPriorities = (selectedNaf: Ref<string>) => {
       .map((p) => ({ ...p, currentPriority: getPriority(p), theme: getTheme(p) }))
   )
 
+  const loadHistory = async () => {
+    try {
+      const data = await $fetch<HistoryEntry[]>('/api/projects/priorities/history')
+      history.value = data.slice().reverse()
+    } catch {
+      // Non bloquant : l'historique est facultatif
+    }
+  }
+
   const refresh = async () => {
     loading.value = true
     errorMessage.value = ''
     successMessage.value = ''
-    modifiedIds.value = new Set()
+    hasChanges.value = false
     try {
       const data = await $fetch<BaserowProject[]>('/api/projects/priorities')
       projects.value = Array.isArray(data) ? data : []
+      // Capture l'état initial pour comparer lors de la sauvegarde
+      originalProjects.value = new Map(
+        projects.value.map((p) => [p.id, { 'Prio': p['Prio'] ?? 9999, 'Prios spécifiques': p['Prios spécifiques'] ?? '' }])
+      )
     } catch {
       errorMessage.value = 'Erreur lors du chargement des projets'
     } finally {
@@ -55,14 +69,11 @@ export const useBaserowPriorities = (selectedNaf: Ref<string>) => {
   const updatePriority = (projectId: number, newVal: string) => {
     const value = parseFloat(newVal)
     if (isNaN(value) || value < 0) return
-
     const project = projects.value.find((p) => p.id === projectId)
     if (!project) return
-
-    modifiedIds.value = new Set([...modifiedIds.value, projectId])
+    hasChanges.value = true
     errorMessage.value = ''
     successMessage.value = ''
-
     const naf = selectedNaf.value
     if (!naf) {
       project['Prio'] = value
@@ -85,22 +96,31 @@ export const useBaserowPriorities = (selectedNaf: Ref<string>) => {
     errorMessage.value = ''
     successMessage.value = ''
     try {
-      const modifiedProjects = projects.value.filter((p) => modifiedIds.value.has(p.id))
-      const updates = modifiedProjects.map((p) => ({
-        id: p.id,
-        Prio: typeof p['Prio'] === 'number' ? p['Prio'] : parseFloat(p['Prio'] as string) || undefined,
-        'Prios spécifiques': p['Prios spécifiques'] || ''
-      }))
-      await $fetch('/api/projects/priorities', {
-        method: 'PATCH',
-        body: { updates, nafCode: selectedNaf.value }
-      })
+      // Seuls les projets réellement modifiés sont envoyés (comparaison avec le snapshot d'origine)
+      const updates = projects.value
+        .filter((p) => {
+          const orig = originalProjects.value.get(p.id)
+          if (!orig) return false
+          const prioChanged = parseFloat(p['Prio'] as string) !== parseFloat(orig['Prio'] as string)
+          const priosSpecChanged = (p['Prios spécifiques'] || '') !== (orig['Prios spécifiques'] || '')
+          return prioChanged || priosSpecChanged
+        })
+        .map((p) => {
+          const orig = originalProjects.value.get(p.id)
+          return {
+            id: p.id,
+            titre: p.Titre || p.title || p.slug || '',
+            oldPrio: parseFloat(orig?.['Prio'] as string) || undefined,
+            Prio: parseFloat(p['Prio'] as string) || undefined,
+            oldPriosSpec: orig?.['Prios spécifiques'] ?? '',
+            'Prios spécifiques': p['Prios spécifiques'] || ''
+          }
+        })
+      await $fetch('/api/projects/priorities', { method: 'PATCH', body: { updates, nafCode: selectedNaf.value } })
       successMessage.value = 'Modifications enregistrées avec succès !'
-      modifiedIds.value = new Set()
-      setTimeout(() => {
-        successMessage.value = ''
-        refresh()
-      }, 1500)
+      hasChanges.value = false
+      await loadHistory()
+      setTimeout(() => { successMessage.value = ''; refresh() }, 1500)
     } catch {
       errorMessage.value = "Erreur lors de l'enregistrement"
     } finally {
@@ -108,5 +128,5 @@ export const useBaserowPriorities = (selectedNaf: Ref<string>) => {
     }
   }
 
-  return { loading, saving, hasChanges, successMessage, errorMessage, sortedProjects, updatePriority, saveChanges, refresh }
+  return { loading, saving, hasChanges, successMessage, errorMessage, sortedProjects, updatePriority, saveChanges, refresh, history, loadHistory }
 }
